requirements.txt
aiosqlite==0.21.0
anyio==4.9.0
APScheduler==3.11.0
certifi==2025.1.31
charset-normalizer==3.4.2
docopt==0.6.2
h11==0.14.0
httpcore==1.0.8
httpx==0.28.1
idna==3.10
pipreqs==0.4.13
pyaes==1.6.1
pyasn1==0.6.1
python-dotenv==1.1.0
python-telegram-bot==22.0
requests==2.32.3
rsa==4.9
sniffio==1.3.1
Telethon==1.39.0
typing_extensions==4.13.2
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.4.0
yarg==0.1.10
